LayoutTests/platform/mac-wk1/fast/loader/redirect-to-invalid-url-using-meta-refresh-disallowed-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - willPerformClientRedirectToURL: http://A=a%25B=b
main frame - didFinishLoadForFrame
main frame - didCancelClientRedirectForFrame
Tests that we do not redirect to an invalid URL initiated by <meta http-equiv="refresh">. This test PASSED if you see an entry in the dumped frame load callbacks of the form: "willPerformClientRedirectToURL: http://A=a%B=b" followed by "didCancelClientRedirectForFrame".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testRunner.didCancelClientRedirect became true
PASS successfullyParsed is true

TEST COMPLETE
